cleanup for release
drop example06 subdir and move to src
fixup pluginConfig.ts fields
drop reference to example06 from rollup.config.js
