+++ package.json
@@ -7,13 +7,11 @@
     "scripts": {
         "build": "rm -rf dist && mkdir dist && SERVE=false rollup -c && cp package.json dist/",
         "start": "rollup -w -c",
-        "example06": "CONFIG=example06 rollup -w -c",
-        "build:all-examples": "for number in 01 02 03 04 05 06 07 08; do CONFIG=example$number SERVE=false rollup -c; done",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/windycom/windy-plugin-template.git"
+        "url": "git+https://github.com/wetterheidi/windy-plugin-foehn-cross-section.git"
     },
     "keywords": [
         "windy",
