+++ src/pluginConfig.ts
@@ -1,16 +1,17 @@
 import type { ExternalPluginConfig } from '@windy/interfaces';
 
 const config: ExternalPluginConfig = {
-    name: 'windy-plugin-my-plugin',
-    version: '0.1.0',
-    icon: '🔌',
-    title: 'This is my first plugin',
-    description: 'This is my first plugin.',
-    author: 'John Doe (optional company name)',
-    repository: 'https://github.com/windycom/windy-plugin-template',
+    name: 'windy-plugin-foehn-chart',
+    version: '0.0.1',
+    title: 'Pressure Difference Charts Alps',
+    icon: '⛰️',
+    description: 'Displays Foehn diagrams for various cross sections.',
+    author: 'Heidi modified IL',
+    repository: 'https://github.com/wetterheidi/windy-plugin-foehn-cross-section',
     desktopUI: 'rhpane',
     mobileUI: 'fullscreen',
-    routerPath: '/my-plugin',
+    routerPath: '/foehn-chart',
+    desktopWidth: 800,
 };
 
 export default config;
